refactor (empty)  과도한 클래스 추출 제거
