+++ src/main/java/kitchenpos/table/dto/EmptyRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.core;
+package kitchenpos.table.dto;
 
 public class EmptyRequest {
